Fixed Correct ID Type In Repository;

+++ Backend/src/main/java/com/coinlibrary/backend/repository/EditionRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface EditionRepository<T, S> extends PagingAndSortingRepository<Edition, Integer>, JpaSpecificationExecutor<Edition> {
     Edition save(Edition edition);
 
-    Optional<Edition> findById(long id);
+    Optional<Edition> findById(int id);
 
     public List<Edition> findByCountry(String country);
